+++ src/js/parsers/mobiledoc/0-2.js
@@ -3,8 +3,8 @@ import {
   MOBILEDOC_IMAGE_SECTION_TYPE,
   MOBILEDOC_LIST_SECTION_TYPE,
   MOBILEDOC_CARD_SECTION_TYPE
-} from '../renderers/mobiledoc';
-import { kvArrayToObject, filter } from "../utils/array-utils";
+} from 'mobiledoc-kit/renderers/mobiledoc/0-2';
+import { kvArrayToObject, filter } from "../../utils/array-utils";
 
 /*
  * Parses from mobiledoc -> post
