+++ src/js/renderers/mobiledoc/index.js
@@ -1,0 +1,16 @@
+import MobiledocRenderer_0_2, { MOBILEDOC_VERSION } from './0-2';
+
+export { MOBILEDOC_VERSION };
+
+export default {
+  render(post, version) {
+    switch (version) {
+      case '0.2.0':
+      case undefined:
+      case null:
+        return MobiledocRenderer_0_2.render(post);
+      default:
+        throw new Error(`Unknown version of mobiledoc renderer requested: ${version}`);
+    }
+  }
+};
