+++ tests/helpers/mobiledoc.js
@@ -1,5 +1,6 @@
 import PostAbstractHelpers from './post-abstract';
-import MobiledocRenderer from 'mobiledoc-kit/renderers/mobiledoc';
+import mobiledocRenderers from 'mobiledoc-kit/renderers/mobiledoc';
+import MobiledocRenderer_0_2 from 'mobiledoc-kit/renderers/mobiledoc/0-2';
 
 /*
  * usage:
@@ -11,8 +12,16 @@ import MobiledocRenderer from 'mobiledoc-kit/renderers/mobiledoc';
  *    })
  *  )
  */
-function build(treeFn) {
-  return MobiledocRenderer.render(PostAbstractHelpers.build(treeFn));
+function build(treeFn, version) {
+  switch (version) {
+    case '0.2.0':
+      return MobiledocRenderer_0_2.render(PostAbstractHelpers.build(treeFn));
+    case undefined:
+    case null:
+      return mobiledocRenderers.render(PostAbstractHelpers.build(treeFn));
+    default:
+      throw new Error(`Unknown version of mobiledoc renderer requested: ${version}`);
+  }
 }
 
 export default {
